Refactor notifications settings
