+++ StudentUsos/Features/Settings/Views/SettingsPage.xaml
@@ -11,35 +11,15 @@
 
     <ScrollView>
         <StackLayout Padding="30" Spacing="10">
-            <Border>
-                <StackLayout>
-                    <Label FontSize="16" HorizontalTextAlignment="Center" Text="{x:Static strings:LocalizedStrings.SettingsPage_CalendarEvents}" />
-                    <Grid ColumnDefinitions="*, auto">
-                        <Label Grid.Column="0" Text="{x:Static strings:LocalizedStrings.Notifications}" />
-                        <Switch Grid.Column="1" x:Name="notificationsSwitch" IsToggled="{Binding AreNotificationsEnabled}" />
-                    </Grid>
-                    <Grid ColumnDefinitions="auto, auto, auto" ColumnSpacing="5" IsVisible="{Binding AreNotificationsEnabled}">
-                        <Label Grid.Column="0" Text="{x:Static strings:LocalizedStrings.Notify}" />
-                        <Entry Grid.Column="1" Keyboard="Numeric" MaxLength="2" Text="{Binding NotificationsDayPicked}" WidthRequest="25">
-                            <Entry.Behaviors>
-                                <behaviors:WholeNumberValidationBehavior />
-                            </Entry.Behaviors>
-                        </Entry>
-                        <Label Grid.Column="2" Text="{x:Static strings:LocalizedStrings.MorePage_DaysBefore}" />
-                    </Grid>
-                    <Grid ColumnDefinitions="auto, auto" ColumnSpacing="5" IsVisible="{Binding AreNotificationsEnabled}">
-                        <Label Grid.Column="0"  Text="{x:Static strings:LocalizedStrings.MorePage_NotifyAt}"></Label>
-                        <TimePicker Grid.Column="1" x:Name="timePicker" Format="HH:mm" Time="{Binding NotificationsTimePicked}" />
-                    </Grid>
-                </StackLayout>
-            </Border>
 
-            <Border >
-                <StackLayout Spacing="10">
-                    <Label FontSize="16" HorizontalTextAlignment="Center" Text="{x:Static strings:LocalizedStrings.Notifications}" />
-                    <Button Text="{x:Static strings:LocalizedStrings.Troubleshooting}" Background="{StaticResource Primary}" 
-                            Clicked="TroubleshootingButton_Clicked"></Button>
-                </StackLayout>
+            <Border Padding="0">
+                <Grid>
+                    <Grid Margin="20" ColumnDefinitions="auto, *" ColumnSpacing="10">
+                        <Image Grid.Column="0" Source="bell" WidthRequest="25"></Image>
+                        <Label Grid.Column="1" Text="{x:Static strings:LocalizedStrings.Notifications}"/>
+                    </Grid>
+                    <Button BackgroundColor="Transparent" Command="{Binding GoToNotificationsSubpageCommand}"/>
+                </Grid>
             </Border>
 
             <Border >
