+++ StudentUsos/Features/Settings/Views/Subpages/EmptyPageTemplate - Copy.xaml
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <controls:CustomContentPageNotAnimated
-    x:Class="StudentUsos.EmptyPageTemplate"
+    x:Class="StudentUsos.Features.Settings.Views.Subpages.EmptyPageTemplate"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:controls="clr-namespace:StudentUsos.Controls"
