+++ StudentUsos/Features/Settings/Views/Subpages/NotificationsSubpage.xaml
@@ -1,0 +1,48 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<controls:CustomContentPageNotAnimated
+    x:Class="StudentUsos.Features.Settings.Views.Subpages.NotificationsSubpage"
+    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:controls="clr-namespace:StudentUsos.Controls"
+    xmlns:strings="clr-namespace:StudentUsos.Resources.LocalizedStrings"
+    xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
+    xmlns:behaviours="clr-namespace:StudentUsos.Behaviours"
+    xmlns:subpages="clr-namespace:StudentUsos.Features.Settings.Views.Subpages"
+    x:DataType="subpages:NotificationsSubpageViewModel">
+
+    <ScrollView>
+        <StackLayout Margin="30">
+
+            <Label FontSize="14" HorizontalTextAlignment="Center"
+                   Text="{x:Static strings:LocalizedStrings.CalendarNotifications}" />
+            <Border Margin="0,5,0,20">
+                <StackLayout>
+                    <Grid ColumnDefinitions="*, auto">
+                        <Label Grid.Column="0" Text="{x:Static strings:LocalizedStrings.Notifications}" />
+                        <Switch Grid.Column="1" IsToggled="{Binding AreNotificationsEnabled}" />
+                    </Grid>
+                    <Grid ColumnDefinitions="auto, auto, auto" ColumnSpacing="5" IsVisible="{Binding AreNotificationsEnabled}">
+                        <Label Grid.Column="0" Text="{x:Static strings:LocalizedStrings.Notify}" />
+                        <Entry Grid.Column="1" Keyboard="Numeric" MaxLength="2" Text="{Binding NotificationsDayPicked}" WidthRequest="25">
+                            <Entry.Behaviors>
+                                <behaviours:WholeNumberValidationBehavior />
+                            </Entry.Behaviors>
+                        </Entry>
+                        <Label Grid.Column="2" Text="{x:Static strings:LocalizedStrings.MorePage_DaysBefore}" />
+                    </Grid>
+                    <Grid ColumnDefinitions="auto, auto" ColumnSpacing="5" IsVisible="{Binding AreNotificationsEnabled}">
+                        <Label Grid.Column="0"  Text="{x:Static strings:LocalizedStrings.MorePage_NotifyAt}"></Label>
+                        <TimePicker Grid.Column="1" Format="HH:mm" Time="{Binding NotificationsTimePicked}" />
+                    </Grid>
+                </StackLayout>
+            </Border>
+
+            <Label FontSize="14" HorizontalTextAlignment="Center" Text="{x:Static strings:LocalizedStrings.PushNotifications}" />
+            <Border Padding="0" Margin="0,5,0,20">
+                    <Button Text="{x:Static strings:LocalizedStrings.Troubleshooting}" Padding="15" Clicked="TroubleshootingButton_Clicked"/>
+            </Border>
+
+        </StackLayout>
+    </ScrollView>
+
+</controls:CustomContentPageNotAnimated>
